refactor(ATT-435): split firmware display.cpp god file into focused units
Extract collaborators from the ~689-line display.cpp into separate translation
units sharing the Display class:
- display_router.cpp   screen routing / transitions
- display_input.cpp    flush + touch input dispatch
- display_overlay.cpp  device-name overlay
- display_popups.cpp   error / insufficient-balance popups

display.cpp retains driver init, lifecycle, and static member definitions.
Behavior preserved (pure code relocation). Bump FIRMWARE_VERSION 1.3.11 -> 1.3.12
for OTA. attractap-touch env builds green.
