+++ apps/attractap/firmware/src/display/display_overlay.cpp
@@ -1,0 +1,77 @@
+#include "display.hpp"
+
+// Persistent device-info overlay drawn on the LVGL top layer: a bottom bar
+// showing the device name and firmware version, present across all screens.
+
+void Display::initDeviceOverlay()
+{
+    lv_obj_t *top = lv_layer_top();
+    lv_obj_remove_flag(top, LV_OBJ_FLAG_SCROLLABLE);
+    // Keep the top layer passive; don't assign a layout directly
+    lv_obj_add_flag(top, LV_OBJ_FLAG_IGNORE_LAYOUT);
+
+    lv_obj_t *deviceInfoContainer = lv_obj_create(top);
+    lv_obj_remove_style_all(deviceInfoContainer);
+    lv_obj_set_width(deviceInfoContainer, lv_pct(100));
+    lv_obj_set_height(deviceInfoContainer, LV_SIZE_CONTENT);
+    lv_obj_set_align(deviceInfoContainer, LV_ALIGN_BOTTOM_MID);
+    lv_obj_set_x(deviceInfoContainer, 0);
+    lv_obj_set_y(deviceInfoContainer, 0);
+    lv_obj_set_flex_flow(deviceInfoContainer, LV_FLEX_FLOW_ROW);
+    // Spread labels horizontally in a bottom bar
+    lv_obj_set_flex_align(deviceInfoContainer, LV_FLEX_ALIGN_SPACE_BETWEEN, LV_FLEX_ALIGN_CENTER, LV_FLEX_ALIGN_START);
+    lv_obj_remove_flag(deviceInfoContainer, LV_OBJ_FLAG_CLICKABLE);
+    lv_obj_remove_flag(deviceInfoContainer, LV_OBJ_FLAG_SCROLLABLE);
+    lv_obj_set_style_pad_left(deviceInfoContainer, 20, LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_pad_right(deviceInfoContainer, 20, LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_pad_top(deviceInfoContainer, 0, LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_pad_bottom(deviceInfoContainer, 10, LV_PART_MAIN | LV_STATE_DEFAULT);
+
+    Display::deviceNameLabel = lv_label_create(deviceInfoContainer);
+    lv_obj_set_width(Display::deviceNameLabel, LV_SIZE_CONTENT);
+    lv_obj_set_height(Display::deviceNameLabel, LV_SIZE_CONTENT);
+    lv_obj_set_x(Display::deviceNameLabel, -75);
+    lv_obj_set_y(Display::deviceNameLabel, -18);
+    lv_obj_set_align(Display::deviceNameLabel, LV_ALIGN_CENTER);
+    lv_label_set_text(Display::deviceNameLabel, Display::deviceNameInitValue.c_str());
+
+    lv_obj_set_style_text_color(Display::deviceNameLabel, lv_color_hex(0xFFFFFF), LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_text_opa(Display::deviceNameLabel, 255, LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_text_font(Display::deviceNameLabel, &lv_font_montserrat_10, LV_PART_MAIN | LV_STATE_DEFAULT);
+
+    lv_obj_t *firmwareLabel = lv_label_create(deviceInfoContainer);
+    lv_obj_set_width(firmwareLabel, LV_SIZE_CONTENT);
+    lv_obj_set_height(firmwareLabel, LV_SIZE_CONTENT);
+    lv_obj_set_align(firmwareLabel, LV_ALIGN_CENTER);
+    lv_label_set_text(firmwareLabel, (String(FIRMWARE_FRIENDLY_NAME) + " v" + String(FIRMWARE_VERSION)).c_str());
+    lv_obj_set_style_text_color(firmwareLabel, lv_color_hex(0xFFFFFF), LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_text_opa(firmwareLabel, 255, LV_PART_MAIN | LV_STATE_DEFAULT);
+    lv_obj_set_style_text_font(firmwareLabel, &lv_font_montserrat_10, LV_PART_MAIN | LV_STATE_DEFAULT);
+}
+
+void Display::setDeviceName(String deviceName)
+{
+    if (!Display::deviceNameLabel)
+    {
+        Display::deviceNameInitValue = deviceName;
+        return;
+    }
+
+    // Schedule label update on LVGL thread to avoid cross-thread access
+    char *text = (char *)malloc(deviceName.length() + 1);
+    if (!text)
+    {
+        return;
+    }
+    strcpy(text, deviceName.c_str());
+    lv_async_call(
+        [](void *p)
+        {
+            if (Display::deviceNameLabel)
+            {
+                lv_label_set_text(Display::deviceNameLabel, (const char *)p);
+            }
+            free(p);
+        },
+        text);
+}
